feat: restructure examples as standalone packages
Each example now lives in its own directory with a package.json that
installs drej from npm (^0.5.0) and a README explaining what it does.
