+++ examples/run-bash-script/README.md
@@ -1,0 +1,17 @@
+# run-bash-script
+
+Run a multi-line bash script inside an isolated sandbox and stream its output.
+
+## Run
+
+```bash
+bun install
+bun start
+```
+
+## What it does
+
+1. Creates an Ubuntu 22.04 sandbox
+2. Uploads and runs a bash script that prints system info, disk usage, and writes/reads a file
+3. Streams output to stdout as it arrives
+4. Deletes the sandbox on completion
