+++ examples/snapshot-replay/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "drej-example-snapshot-replay",
+  "private": true,
+  "version": "0.0.0",
+  "scripts": {
+    "start": "bun index.ts"
+  },
+  "dependencies": {
+    "drej": "^0.5.0"
+  }
+}
